removed warnings
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@8556 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/microsim/MSEdge.cpp
@@ -257,7 +257,7 @@ MSEdge::getDepartLane(const MSVehicle &v) const throw() {
     }
     switch (pars.departLaneProcedure) {
     case DEPART_LANE_GIVEN:
-        if (myLanes->size() <= pars.departLane || !(*myLanes)[pars.departLane]->allowsVehicleClass(v.getVehicleType().getVehicleClass())) {
+        if ((int) myLanes->size() <= pars.departLane || !(*myLanes)[pars.departLane]->allowsVehicleClass(v.getVehicleType().getVehicleClass())) {
             return 0;
         }
         return (*myLanes)[pars.departLane];
